[GUI] re-group tabs
