+++ gui/src/main/java/org/astraea/gui/Main.java
@@ -18,27 +18,19 @@ package org.astraea.gui;
 
 import java.util.List;
 import javafx.application.Application;
+import javafx.geometry.Side;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 import org.astraea.gui.pane.TabPane;
 import org.astraea.gui.tab.AboutTab;
 import org.astraea.gui.tab.BalancerTab;
 import org.astraea.gui.tab.BrokerTab;
-import org.astraea.gui.tab.ConfigTab;
-import org.astraea.gui.tab.ConsumerTab;
-import org.astraea.gui.tab.CreateTopicTab;
-import org.astraea.gui.tab.MetricsTab;
-import org.astraea.gui.tab.MovingReplicaTab;
+import org.astraea.gui.tab.ClientTab;
 import org.astraea.gui.tab.PartitionTab;
-import org.astraea.gui.tab.ProducerTab;
 import org.astraea.gui.tab.QuotaTab;
+import org.astraea.gui.tab.ReplicaTab;
 import org.astraea.gui.tab.SettingTab;
 import org.astraea.gui.tab.TopicTab;
-import org.astraea.gui.tab.TransactionTab;
-import org.astraea.gui.tab.UpdateBrokerTab;
-import org.astraea.gui.tab.UpdatePartitionTab;
-import org.astraea.gui.tab.UpdateQuotaTab;
-import org.astraea.gui.tab.UpdateTopicTab;
 
 /**
  * Since the Java launcher checks if the main class extends javafx.application.Application, and in
@@ -63,23 +55,15 @@ public class Main {
       stage.setScene(
           new Scene(
               TabPane.of(
+                  Side.BOTTOM,
                   List.of(
                       SettingTab.of(context),
                       BrokerTab.of(context),
-                      MetricsTab.of(context),
                       TopicTab.of(context),
                       PartitionTab.of(context),
-                      ConfigTab.of(context),
-                      ConsumerTab.of(context),
-                      ProducerTab.of(context),
-                      TransactionTab.of(context),
-                      MovingReplicaTab.of(context),
+                      ReplicaTab.of(context),
+                      ClientTab.of(context),
                       QuotaTab.of(context),
-                      CreateTopicTab.of(context),
-                      UpdateBrokerTab.of(context),
-                      UpdateTopicTab.of(context),
-                      UpdatePartitionTab.of(context),
-                      UpdateQuotaTab.of(context),
                       BalancerTab.of(context),
                       AboutTab.of(context))),
               300,
