+++ gui/src/main/java/org/astraea/gui/tab/TopicTab.java
@@ -16,30 +16,128 @@
  */
 package org.astraea.gui.tab;
 
-import java.util.Arrays;
+import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionStage;
 import java.util.stream.Collectors;
+import javafx.geometry.Side;
 import org.astraea.common.DataSize;
 import org.astraea.common.LinkedHashMap;
 import org.astraea.common.Utils;
 import org.astraea.common.admin.Broker;
 import org.astraea.common.admin.NodeInfo;
 import org.astraea.common.admin.Partition;
+import org.astraea.common.admin.Topic;
+import org.astraea.common.admin.TopicConfigs;
 import org.astraea.common.metrics.broker.HasRate;
 import org.astraea.common.metrics.broker.ServerMetrics;
 import org.astraea.gui.Context;
+import org.astraea.gui.pane.BorderPane;
 import org.astraea.gui.pane.PaneBuilder;
 import org.astraea.gui.pane.Tab;
+import org.astraea.gui.pane.TabPane;
 
 public class TopicTab {
-  private static List<Map<String, Object>> result(
-      List<Partition> partitions, List<Broker> nodes, Map<String, Map<String, Object>> metrics) {
+
+  private static Tab metricsTab(Context context) {
+    return Tab.of(
+        "metrics",
+        PaneBuilder.of()
+            .radioButtons(ServerMetrics.Topic.values())
+            .searchField("topic name")
+            .buttonAction(
+                (input, logger) ->
+                    CompletableFuture.supplyAsync(
+                        () -> {
+                          var metric =
+                              input
+                                  .selectedRadio()
+                                  .map(o -> (ServerMetrics.Topic) o)
+                                  .orElse(ServerMetrics.Topic.BYTES_IN_PER_SEC);
+                          var nodeMeters =
+                              context.clients().entrySet().stream()
+                                  .collect(
+                                      Utils.toSortedMap(
+                                          Map.Entry::getKey,
+                                          entry ->
+                                              metric.fetch(entry.getValue()).stream()
+                                                  .filter(m -> input.matchSearch(m.topic()))
+                                                  .collect(Collectors.toList())));
+                          var topics =
+                              nodeMeters.values().stream()
+                                  .flatMap(Collection::stream)
+                                  .map(ServerMetrics.Topic.Meter::topic)
+                                  .collect(Collectors.toSet());
+
+                          return topics.stream()
+                              .map(
+                                  topic -> {
+                                    var map = new LinkedHashMap<String, Object>();
+                                    map.put("topic", topic);
+                                    nodeMeters.forEach(
+                                        (nodeInfo, meters) -> {
+                                          var key = String.valueOf(nodeInfo.id());
+                                          var value =
+                                              meters.stream()
+                                                  .filter(m -> m.topic().equals(topic))
+                                                  .mapToDouble(HasRate::fiveMinuteRate)
+                                                  .sum();
+                                          switch (metric) {
+                                            case BYTES_IN_PER_SEC:
+                                            case BYTES_OUT_PER_SEC:
+                                              map.put(key, DataSize.Byte.of((long) value));
+                                              break;
+                                            default:
+                                              map.put(key, value);
+                                              break;
+                                          }
+                                        });
+                                    return map;
+                                  })
+                              .collect(Collectors.toList());
+                        }))
+            .build());
+  }
+
+  private static Tab configTab(Context context) {
+    return Tab.of(
+        "config",
+        PaneBuilder.of()
+            .searchField("config key")
+            .buttonAction(
+                (input, logger) ->
+                    context
+                        .admin()
+                        .topicNames(true)
+                        .thenCompose(context.admin()::topics)
+                        .thenApply(
+                            topics -> topics.stream().map(t -> Map.entry(t.name(), t.config())))
+                        .thenApply(
+                            items ->
+                                items
+                                    .map(
+                                        e -> {
+                                          var map = new LinkedHashMap<String, Object>();
+                                          map.put("topic", e.getKey());
+                                          e.getValue().raw().entrySet().stream()
+                                              .filter(entry -> input.matchSearch(entry.getKey()))
+                                              .sorted(Map.Entry.comparingByKey())
+                                              .forEach(
+                                                  entry ->
+                                                      map.put(entry.getKey(), entry.getValue()));
+                                          return map;
+                                        })
+                                    .collect(Collectors.toList())))
+            .build());
+  }
+
+  private static List<Map<String, Object>> basicResult(
+      List<Partition> partitions, List<Broker> brokers) {
     var topicSize =
-        nodes.stream()
+        brokers.stream()
             .flatMap(n -> n.folders().stream().flatMap(d -> d.partitionSizes().entrySet().stream()))
             .collect(Collectors.groupingBy(e -> e.getKey().topic()))
             .entrySet()
@@ -71,7 +169,6 @@ public class TopicTab {
                           .mapToLong(Partition::maxTimestamp)
                           .max()
                           .orElse(-1L)));
-              result.putAll(metrics.getOrDefault(topic, Map.of()));
               tps.get(topic).stream()
                   .flatMap(p -> p.replicas().stream())
                   .collect(Collectors.groupingBy(NodeInfo::id))
@@ -85,53 +182,9 @@ public class TopicTab {
         .collect(Collectors.toList());
   }
 
-  private static CompletionStage<Map<String, Map<String, Object>>> topicMetrics(Context context) {
-    if (!context.hasMetrics()) return CompletableFuture.completedFuture(Map.of());
-    return CompletableFuture.supplyAsync(
-            () ->
-                context.clients().values().stream()
-                    .flatMap(
-                        client ->
-                            Arrays.stream(ServerMetrics.Topic.values())
-                                .flatMap(
-                                    m ->
-                                        MetricsTab.tryToFetch(() -> m.fetch(client).stream())
-                                            .stream()))
-                    .flatMap(m -> m)
-                    .collect(Collectors.groupingBy(ServerMetrics.Topic.Meter::topic)))
-        .thenApply(
-            ms ->
-                ms.entrySet().stream()
-                    .collect(
-                        Collectors.toMap(
-                            Map.Entry::getKey,
-                            e ->
-                                e.getValue().stream()
-                                    .collect(Collectors.groupingBy(ServerMetrics.Topic.Meter::type))
-                                    .entrySet()
-                                    .stream()
-                                    .collect(
-                                        Utils.toSortedMap(
-                                            e2 -> e2.getKey().alias(),
-                                            e2 -> {
-                                              switch (e2.getKey()) {
-                                                case BYTES_IN_PER_SEC:
-                                                case BYTES_OUT_PER_SEC:
-                                                  return DataSize.Byte.of(
-                                                      (long)
-                                                          e2.getValue().stream()
-                                                              .mapToDouble(HasRate::fiveMinuteRate)
-                                                              .sum());
-                                                default:
-                                                  return e2.getValue().stream()
-                                                      .mapToDouble(HasRate::fiveMinuteRate)
-                                                      .sum();
-                                              }
-                                            })))));
-  }
-
-  public static Tab of(Context context) {
-    var pane =
+  private static Tab basicTab(Context context) {
+    return Tab.of(
+        "basic",
         PaneBuilder.of()
             .searchField("topic name")
             .buttonAction(
@@ -154,11 +207,140 @@ public class TopicTab {
                                             context
                                                 .admin()
                                                 .brokers()
-                                                .thenCombine(
-                                                    topicMetrics(context),
-                                                    (brokers, metrics) ->
-                                                        result(partitions, brokers, metrics)))))
-            .build();
-    return Tab.of("topic", pane);
+                                                .thenApply(
+                                                    brokers -> basicResult(partitions, brokers)))))
+            .build());
+  }
+
+  public static Tab alterTab(Context context) {
+    var numberOfPartitions = "number of partitions";
+    return Tab.dynamic(
+        "alter",
+        () ->
+            context
+                .admin()
+                .topicNames(false)
+                .thenCompose(context.admin()::topics)
+                .thenApply(
+                    topics ->
+                        BorderPane.selectableTop(
+                            topics.stream()
+                                .collect(
+                                    Utils.toSortedMap(
+                                        Topic::name,
+                                        topic ->
+                                            PaneBuilder.of()
+                                                .buttonName("ALTER")
+                                                .input(
+                                                    numberOfPartitions,
+                                                    false,
+                                                    true,
+                                                    String.valueOf(topic.topicPartitions().size()))
+                                                .input(
+                                                    TopicConfigs.DYNAMICAL_CONFIGS.stream()
+                                                        .collect(
+                                                            Collectors.toMap(
+                                                                k -> k,
+                                                                k ->
+                                                                    topic
+                                                                        .config()
+                                                                        .value(k)
+                                                                        .orElse(""))))
+                                                .buttonListener(
+                                                    (input, logger) -> {
+                                                      var allConfigs =
+                                                          new HashMap<>(input.nonEmptyTexts());
+                                                      var partitions =
+                                                          Integer.parseInt(
+                                                              allConfigs.remove(
+                                                                  numberOfPartitions));
+                                                      return context
+                                                          .admin()
+                                                          .setConfigs(topic.name(), allConfigs)
+                                                          .thenCompose(
+                                                              ignored ->
+                                                                  context
+                                                                      .admin()
+                                                                      .unsetConfigs(
+                                                                          topic.name(),
+                                                                          input.emptyValueKeys()))
+                                                          .thenCompose(
+                                                              ignored ->
+                                                                  partitions
+                                                                          == topic
+                                                                              .topicPartitions()
+                                                                              .size()
+                                                                      ? CompletableFuture
+                                                                          .completedFuture(null)
+                                                                      : context
+                                                                          .admin()
+                                                                          .addPartitions(
+                                                                              topic.name(),
+                                                                              partitions))
+                                                          .thenAccept(
+                                                              ignored ->
+                                                                  logger.log(
+                                                                      "succeed to alter "
+                                                                          + topic.name()));
+                                                    })
+                                                .build())))));
+  }
+
+  public static Tab createTab(Context context) {
+    var topicNameKey = "topic";
+    var numberOfPartitionsKey = "number of partitions";
+    var numberOfReplicasKey = "number of replicas";
+    return Tab.of(
+        "create",
+        PaneBuilder.of()
+            .buttonName("CREATE")
+            .input(topicNameKey, true, false)
+            .input(numberOfPartitionsKey, false, true)
+            .input(numberOfReplicasKey, false, true)
+            .input(TopicConfigs.ALL_CONFIGS)
+            .buttonListener(
+                (input, logger) -> {
+                  var allConfigs = new HashMap<>(input.nonEmptyTexts());
+                  var name = allConfigs.remove(topicNameKey);
+                  return context
+                      .admin()
+                      .topicNames(true)
+                      .thenCompose(
+                          names -> {
+                            if (names.contains(name))
+                              return CompletableFuture.failedFuture(
+                                  new IllegalArgumentException(name + " is already existent"));
+
+                            return context
+                                .admin()
+                                .creator()
+                                .topic(name)
+                                .numberOfPartitions(
+                                    Optional.ofNullable(allConfigs.remove(numberOfPartitionsKey))
+                                        .map(Integer::parseInt)
+                                        .orElse(1))
+                                .numberOfReplicas(
+                                    Optional.ofNullable(allConfigs.remove(numberOfReplicasKey))
+                                        .map(Short::parseShort)
+                                        .orElse((short) 1))
+                                .configs(allConfigs)
+                                .run()
+                                .thenAccept(i -> logger.log("succeed to create topic:" + name));
+                          });
+                })
+            .build());
+  }
+
+  public static Tab of(Context context) {
+    return Tab.of(
+        "topic",
+        TabPane.of(
+            Side.TOP,
+            List.of(
+                basicTab(context),
+                configTab(context),
+                metricsTab(context),
+                createTab(context),
+                alterTab(context))));
   }
 }
